project-3/requirements.txt
requests~=2.32.3
mysql-connector-python~=9.2.0
Flask~=3.1.0